extracted image creation function from trails controller to utils
