+++ .vscode/settings.json
@@ -1,3 +1,3 @@
 {
-  "cSpell.words": ["Flothemes", "originalname", "pois"]
+  "cSpell.words": ["Flothemes", "originalname", "Phusion", "pois"]
 }
