+++ server/controllers/trailControllers.js
@@ -1,24 +1,17 @@
 const { body, validationResult } = require("express-validator");
-const sharp = require("sharp");
 
 const { loggers } = require("winston");
 const logger = loggers.get("logger");
 
-const {
-  SAVE_DIRECTORY,
-  QUALITY: quality,
-  RESIZE,
-  VALID_IMAGE_TYPES,
-} = require("../config/imageUpload");
-const { distance } = require("../utils/distance");
-const { ensureDirExists, getFileName } = require("../utils/fileHelpers");
-
 const { Trail, TrailCoords, User, PointsOfInterest } = require("../models");
+
+const { VALID_IMAGE_TYPES } = require("../config/imageUpload");
+const { distance } = require("../utils/distance");
 const {
-  validationErrors,
   checkLengthOfObjectArrays,
   makeObjectArray,
 } = require("../utils/helpers");
+const { getImageLinks } = require("../utils/images");
 
 // returns a list of all of the trails and the trail's points
 exports.listTrails = async (req, res) => {
@@ -275,49 +268,3 @@ exports.saveTrail = [
     }
   },
 ];
-
-/**
- * Takes in a trailId, file (from req.files) and a type, either POI or Hazard to define where the image will be stored. Images will be stored in:
- *
- *    ./public/images/<trailId>/<POI | Hazard>/
- *
- * @param {number} trailId
- * @param {fileObject} file
- * @param {"POI" | "Hazard"} type
- * @returns The link to the files storage location
- */
-async function getImageLinks(trailId, file, type) {
-  try {
-    // ensure filesystem exists for save
-    const path = SAVE_DIRECTORY + trailId + "/POI/";
-    // console.log("getImageLinks: path", path);
-
-    await ensureDirExists(path);
-
-    const { buffer, originalname } = file;
-    const { fileName } = getFileName(originalname);
-    const link = `${path}${fileName}.webp`;
-
-    await sharp(buffer)
-      .resize(RESIZE)
-      .webp({ quality })
-      .toFile(link)
-      .catch((err) => {
-        console.log(err);
-        logger.debug(err, {
-          controller: "getImageLinks",
-          errorMsg: "getImageLinks Sharp Error writing file",
-        });
-        throw new Error(err);
-      });
-
-    // console.log("getImageLinks: link", link);
-    return link;
-  } catch (err) {
-    console.log(err);
-    logger.debug(err, {
-      controller: "getImageLinks",
-      errorMsg: "catch error",
-    });
-  }
-}
